+++ distributors.cabal
@@ -48,6 +48,8 @@ library
       Data.Profunctor.Grammar.Parsector
       Data.Profunctor.Monadic
       Data.Profunctor.Monoidal
+      Data.Profunctor.Separator
+      Data.Traversable.Homogeneous
   other-modules:
       Paths_distributors
   autogen-modules:
