+++ src/Control/Lens/Bifocal.hs
@@ -39,6 +39,7 @@ import Control.Lens.PartialIso
 import Data.Profunctor
 import Data.Profunctor.Distributor
 import Data.Profunctor.Filtrator
+import Data.Profunctor.Separator
 import Witherable
 
 {- | `Bifocal`s are bidirectional parser optics.
