+++ test/Examples/Lambda.hs
@@ -12,6 +12,7 @@ import Control.Lens.Grammar.Token
 import Control.Lens.PartialIso
 import Data.Profunctor.Distributor
 import Data.Profunctor.Monoidal
+import Data.Profunctor.Separator
 
 -- | Abstract syntax tree for lambda calculus terms
 data Lambda
