+++ test/Examples/SemVer.hs
@@ -14,6 +14,7 @@ import Control.Lens.PartialIso
 import Data.Profunctor.Distributor
 import qualified Data.Profunctor.Monadic as P
 import Data.Profunctor.Monoidal
+import Data.Profunctor.Separator
 import Numeric.Natural
 
 -- | Semantic version structure following semver.org specification
